refactor: organize imports and files on ordersTable module
