+++ apps/cowswap-frontend/src/modules/ordersTable/const/tabs.ts
@@ -20,3 +20,6 @@ export const HISTORY_TAB: OrderTab = {
 export const ORDERS_TABLE_TABS: OrderTab[] = [OPEN_TAB, HISTORY_TAB]
 
 export const ORDERS_TABLE_PAGE_SIZE = 10
+
+export const ORDERS_TABLE_TAB_KEY = 'tab'
+export const ORDERS_TABLE_PAGE_KEY = 'page'
