+++ apps/cowswap-frontend/src/modules/ordersTable/index.tsx
@@ -1,1 +1,6 @@
+export * from './const/tabs'
 export * from './containers/OrdersTableWidget'
+export * from './hooks/useGetBuildOrdersTableUrl'
+export * from './hooks/useNavigateToOpenOrdersTable'
+export * from './types'
+export * from './utils/buildOrdersTableUrl'
