+++ apps/cowswap-frontend/src/modules/ordersTable/types.ts
@@ -1,0 +1,9 @@
+export interface OrdersTablePageParams {
+  tabId: string
+  pageNumber: number
+}
+
+export enum TabOrderTypes {
+  LIMIT = 'limit',
+  ADVANCED = 'advanced',
+}
